feat(setup,switch,workflows): remove BMAD integration and the disabledSkills mechanism
Reverses #41/#42. BMAD was uninstalled from the environment; its coexistence
integration and the disabledSkills mechanism (which existed only for BMAD routing)
were permanently-dead code threaded through 13 live skill files.

- switch: drop Step 5b "BMAD Detection Check"; renumber remaining checks
- setup: drop Step 2b (BMAD detection + skill-preference UI); finalize no longer
  writes disabledSkills/bmadDetected; critiqueMethod kept (reflexion is now the
  only supported value; the bmad-party-mode option is gone)
- 12 workflow skills: remove the disabledSkills check from the config-loading
  preamble; scrub BMAD from skill descriptions and critiqueMethod checks
- collapse the config-loading shared block to one file (the standard/headless
  variants differed only by the disabled-skill headless-cleanup paragraph);
  update scripts/sync_shared_blocks.py manifest + the drift guard
- delete templates/CLAUDE-bmad-routing.md and tests/hooks/test_bmad_detection.py
  (its non-integration lints -- headless protocol, mandatory-steps, critiqueMethod
  presence -- moved to tests/hooks/test_headless.py)
- docs/config-reference.md + README: drop bmadDetected/disabledSkills schema and
  the BMAD section, reword changelog; bump plugin 2.44.0 -> 2.45.0

Graceful compat: leftover bmadDetected/disabledSkills fields in existing workspace
JSONs are ignored -- no hook reads or validates them.

pytest: 1415 -> 1384 passed (31 BMAD-only tests removed; 22 non-BMAD lints preserved).

Closes #123
